backend/bank/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-otp==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
